user requirements.yml for miniconda in docker

+++ docker/user_python-ubuntu.sh
@@ -52,14 +52,19 @@ miniconda)
 esac
 
 if [ "$PYTHON" = "miniconda" ]; then
-  if [ -f requirements_conda_forge.txt ]; then
-    conda install --yes -c conda-forge -c intel -c ccpi -c astra-toolbox --file requirements_conda_forge.txt
+  if [ -f requirements.yml ]; then
+    conda env update --file requirements.yml
+  else
+  # Python (runtime)
+    if [ -f requirements.txt ]; then
+      pip install -U -r requirements.txt
+    fi
   fi
   conda update -c conda-forge -y --all
   conda clean -y --all
-fi
-
-# Python (runtime)
-if [ -f requirements.txt ]; then
-  pip install -U -r requirements.txt
+else
+  # Python (runtime)
+  if [ -f requirements.txt ]; then
+    pip install -U -r requirements.txt
+  fi
 fi
